pyproject.toml
[project]
name = "cafe-order"
version = "0.1.0"
description = "Cafe order management system."
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.12, <3.13"
dependencies = [
    "django>=5.1.5",
    "django-environ>=0.12.0",
    "djangorestframework>=3.15.2",
    "drf-spectacular>=0.28.0",
    "drf-spectacular-sidecar>=2024.12.1",
]

[dependency-groups]
dev = [
    "djlint>=1.36.4",
    "factory-boy>=3.3.1",
    "pytest>=8.3.4",
    "pytest-django>=4.9.0",
    "ruff>=0.9.2",
]

[tool.djlint]
indent = "2"
max_line_length = "79"
preserve_blank_lines = true
profile = "django"

[tool.pytest.ini_options]
pythonpath = "cafe_order/"
DJANGO_SETTINGS_MODULE = "cafe_order.settings"
norecursedirs = [".venv/*"]
addopts = [
    "--import-mode=importlib",
    "-vv",
]

[tool.ruff]
line-length = 79
target-version = "py312"
extend-exclude = ["migrations"]

[tool.ruff.format]
quote-style = "single"

[tool.ruff.lint]
select = [
    "E",     # pycodestyle errors     https://docs.astral.sh/ruff/rules/#error-e
    "W",     # pycodestyle warnings   https://docs.astral.sh/ruff/rules/#warning-w
    "C90",   # mccabe                 https://docs.astral.sh/ruff/rules/#mccabe-c90
    "N",     # pep8-naming            https://docs.astral.sh/ruff/rules/#pep8-naming-n
    "D",     # pydocstyle             https://docs.astral.sh/ruff/rules/#pydocstyle-d
    "I",     # isort                  https://docs.astral.sh/ruff/rules/#isort-i
    "F",     # pyflakes               https://docs.astral.sh/ruff/rules/#pyflakes-f
    "C4",    # flake8-comprehensions  https://docs.astral.sh/ruff/rules/#flake8-comprehensions-c4
    "B",     # flake8-bugbear         https://docs.astral.sh/ruff/rules/#flake8-bugbear-b
    "Q",     # flake8-quotes          https://docs.astral.sh/ruff/rules/#flake8-quotes-q
    "T20",   # flake8-print           https://docs.astral.sh/ruff/rules/#flake8-print-t20
    "SIM",   # flake8-simplify        https://docs.astral.sh/ruff/rules/#flake8-simplify-sim
    "PTH",   # flake8-use-pathlib     https://docs.astral.sh/ruff/rules/#flake8-use-pathlib-pth
    "TCH",   # flake8-type-checking   https://docs.astral.sh/ruff/rules/#flake8-type-checking-tch
    "PERF",  # Perflint               https://docs.astral.sh/ruff/rules/#perflint-perf
    "UP",    # pyupgrade              https://docs.astral.sh/ruff/rules/#pyupgrade-up
]
ignore = [
    "D100",     # missing docstring in public module
    "D101",
    "D102",
    "D103",
    "D104",     # missing docstring in public package
    "D105",     # missing docstring in magic methods
    "D106",     # missing docstring in public nested class
    "D107",     # missing docstring in __init__
]

[tool.ruff.lint.flake8-quotes]
inline-quotes = "single"

[tool.ruff.lint.per-file-ignores]
"settings.py" = ["E501"]

[tool.ruff.lint.pydocstyle]
convention = "google"
